Fix icdm 2.1 test case without uat
